post task

+++ entity/user.go
@@ -1,7 +1,7 @@
 package entity
 
 type User struct {
-	ID    int    `json:"ID,omitempty"`
+	ID    int64  `json:"ID,omitempty"`
 	Nome  string `json:"Nome,omitempty"`
 	Email string `json:"Email,omitempty"`
 	Senha string `json:"Senha,omitempty"`
